Resolve Nextflow race condition (#19)
Until this commit, OA brazenly wrote a Nextflow config file to the outward assembly repo for each invocation of OA with Batch. This meant an assembly run's Nextflow config wasn't found in the assembly work dir; worse, there could be race conditions or conflicts between multiple assembly runs. This commit refactors how we pass parameters to Nextflow and in doing so does part of the work of upgrading to NF25.